requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.1
google-auth==2.34.0
h11==0.14.0
httptools==0.6.1
idna==3.7
Jinja2==3.1.4
kubernetes==25.3.0
MarkupSafe==2.1.5
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websocket-client==1.8.0
websockets==12.0
